REG, MAP, large.
Finished up with the Interactive Map and Registration. Added like a
million methods in the Registration are, might want to look at
refactoring that somehow. The point and click methods of the
WrappedElement class now return the instance, makes navigation easier.
Added a complement to wait_until_gone: wait_until_present.