Right, Never Mind Email.
Email may not be possible after all, there doesn't seem to be any
obvious way to determine which email a given email is, nor is F5-spam
something that IMAP is too good at. Also, finished refactoring.

That's not to say that it works, it almost certainly has major issues
somewhere, but at least the first pass is finished.