CN DUN
FINISHED CHINA MODE.
With possible exception of the TRN test, which takes like an hour to
run, so I'll do that in the jumpbox or something?
Added another button duplicate for CNmod2, added a method so that the
nav menu doesn't keep getting in front of things, and a whole bunch of #
China doesn't have this.'s.