Starting on Backups
Trying out a system to add redundancy to the tests.
Basically, if something fails, instead of stopping entirely, use a
workaround and see if anything behind it is also broken.
Then fail the test, of course.

Also added a couple of skipTests to the script where it is not
applicable, instead of marking it as passed.