Significant Contingencies
Did about half of the tests' backup routes.
Up to the Training Summary test. That'll be tricky.
Removed the Fauxlement, selene will now handle all of the alternate
execution, I guess. Added a backup static class though, containing the
urls of all of the pages that can be accessed if the nav header happens
to be down.