Documented Assertion Failures
Realising that the assert calls don't actually give the full information
once collected in that DR errors list thing, I have now added
human-readable error messages to each of the calls.

Also, the Queensland and Victoria modules, I somehow listed them under
Core.

So.

Went and fixed that one.

Also added a backup for the login not redirecting to a /secure page.