Refactor
After changing the attach_link method to take a list of names instead of
calling it one-at-a-time,
added a bunch more invocations of it.

Changed documentation of switch_to_window to reflect that it does
actually take an index to refer with.

And, added some more progress on AUS. Not a lot though/