Fixes More
Changed globs to a dict, removed Nict entirely.

Made modulescripts into a bit of a properties file.

Refactored authentication out of the urls, added some workaround for
dealing with browser drivers that can't use Basic Authentication.

Fixed the module element scrolling. Super really for real this time.

And, implemented a bunch of the actual singletonians.