Changes
Added a fix to the Training Summary tet that doesn't actually work, you
need to actually navigate to the last slide, not just jump there, so.

Corrected the Backup url for the Assignments page.

Added some actual stuff to fenester. Doesn't display results or initiate
testing yet, but it does read and populate half the stuff from the
config file now.

Added a Fancy Langs odict to modulescripts, used in fenester, but this
file seems like the best place to put locale lists.

Rearranged selene to allow for external test initiation.