Maybe Module Fixes
Commented out fenestration for now.

Added a thing to the modulejson reader, now shouldn't break if it
encounters test or new modules.

Added a bit of an introduction to the Results Displayator. Should help
read the results a bit.

Messed with the Training Summary Tests and Components a little, trying
to get it to at least not break things when it breaks. No luck so far,
going to have to retire MUDA, as it causes a /en emailset.