TRN Fix
Added a thing for testing the testing: a function that sets up a globals
dict and driver with a minimum of fuss.

Fixed Training Summary Test, finally. Turns out, don't need to have that
wait_for_module call right after a method that already calls it.

Also, fixed switch_to_frame(None), added an else after that if.