Chromatic Users, Updated Mufflers
Changed the USERS global in modulescripts, now it's USER, just the one
username, but set it up to have access to all locales' assigments.
Should be easy enough for CN anyway, with just the one locale there
anyway.
Added a proper PASSWORD field there too, no more hardcoding that.

Updated a bunch of references to pylint:disable-msg, changed those to
pylint:disable, as it's been deprecated?

Also added a determined click to the Components and Driver, just in case
you don't want to find a new element when clicking on something that's
behind a screen.